requirements.txt
google-generativeai==0.8.4
fastapi==0.115.12
uvicorn[standard]
uvicorn==0.34.0
pydantic==2.11.1
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
pytest==8.3.5
black==25.1.0
flake8==7.2.0 
httpx
